Preferential match included strings
It can be tricky, with the current fuzzy matching, to do non-fuzzy
sorting. For example, searching a bank of components for 'icon'
will find many results, very few of which are the ones you want.

Preferential treatment was given for finding strings that started
with a given string; this PR extends that preference for any
substring, regardless of its position in the target.